Add data preprocessing for semantic segmentation
