+++ automotive/semantic-segmentation/user.conf
@@ -2,3 +2,4 @@
 # The key has the format 'model.scenario.key'. Value is mostly int64_t.
 # Model maybe '*' as wildcard. In that case the value applies to all models.
 # All times are in milli seconds
+*.*.performance_sample_count_override = 256
